model-driven-telemetry/netconf/pyproject.toml
[tool.poetry]
name = "netconf_dial_in"
version = "0.1.0"
description = ""
authors = ["Dmitry Figol <git@dmfigol.me>"]

[tool.poetry.dependencies]
python = "^3.7"
ncclient = { git = "https://github.com/CiscoDevNet/ncclient.git", branch = "master" }
xmltodict = "^0.12.0"
lxml = "^4.5.0"
jxmlease = "^1.0.1"
kafka-python = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"